+++ jeap-process-archive-config-repository/src/main/java/ch/admin/bit/jeap/processarchive/configuration/json/model/DomainEventArchiveConfiguration.java
@@ -10,7 +10,7 @@ public class DomainEventArchiveConfiguration {
     private String clusterName;
     private String domainEventArchiveDataProvider;
     private String referenceProvider;
-    private String messageProvider;
+    private String archiveDataReferenceProvider;
     private String condition;
     private String uri;
     private String oauthClientId;