+++ jeap-process-archive-config-repository/src/test/java/ch/admin/bit/jeap/processarchive/configuration/json/test/TestArchiveDataReferenceProvider.java
@@ -2,9 +2,10 @@ package ch.admin.bit.jeap.processarchive.configuration.json.test;
 
 import ch.admin.bit.jeap.messaging.model.Message;
 import ch.admin.bit.jeap.processarchive.plugin.api.archivedata.ArchiveDataReference;
-import ch.admin.bit.jeap.processarchive.plugin.api.archivedata.MessageProvider;
+import ch.admin.bit.jeap.processarchive.plugin.api.archivedata.ArchiveDataReferenceProvider;
 
-public class TestMessageProvider implements MessageProvider<Message> {
+public class TestArchiveDataReferenceProvider implements ArchiveDataReferenceProvider<Message> {
+
     @Override
     public ArchiveDataReference getReference(Message message) {
         throw new UnsupportedOperationException();